+++ vivarium/experiments/large_database_experiment.py
@@ -1,5 +1,5 @@
 import random
-from vivarium.core.experiment import Experiment
+from vivarium.core.engine import Engine
 from vivarium.core.process import Process
 from vivarium.core.composer import Composer
 from vivarium.core.emitter import (
@@ -65,7 +65,7 @@ def run_large_initial_emit():
         'emitter': 'database'
     }
 
-    experiment = Experiment({
+    experiment = Engine({
         'processes': composite['processes'],
         'topology': composite['topology'],
         **settings})
